pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "airplay-receiver"
version = "11.0.0"
description = "AirPlay receiver for Windows and Linux with Music Assistant support"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["airplay", "raop", "music", "audio", "receiver"]
dependencies = [
    "sounddevice>=0.4.6",
    "numpy>=1.24",
    "zeroconf>=0.115",
    "pystray>=0.19",
    "Pillow>=10.0",
    "av>=12.0.0",
    "platformdirs>=4.0",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.0",
    "bandit[toml]>=1.7",
    "safety>=3.0",
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
airplay-receiver = "airplay_receiver.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.bandit]
exclude_dirs = ["tests", ".github"]
skips = ["B311"]   # random for non-crypto use

[tool.pytest.ini_options]
testpaths = ["tests"]
